Update Weather Alert

+++ app/src/main/java/com/example/weatherforecast/model/local/AlarmData.kt
@@ -4,7 +4,12 @@ import androidx.room.Entity
 import androidx.room.PrimaryKey
 
 @Entity(tableName ="AlarmData")
-data class AlarmData (val alarmStatus:Boolean,val time:Long,val forNext:String,val alarmSound:String,val event:String){
+data class AlarmData(
+    val alarmStatus:Boolean,
+    val time: Double,
+    val forNext:String,
+    val alarmSound:String,
+    val event:String){
   @PrimaryKey(autoGenerate = true)
     var id:Int=0
 }
